requirements.txt
pydantic
opencv-python
numpy
uvicorn[standard]
fastapi
ultralytics 
pyorbbecsdk
pynetworktables